Separate MachineListTable from MachineList.
