+++ ui/src/app/machines/views/MachineList/MachineListTable/CoresColumn/CoresColumn.test.js
