+++ ui/src/app/machines/views/MachineList/MachineListTable/CoresColumn/index.js
