+++ ui/src/app/machines/views/MachineList/MachineListTable/DisksColumn/DisksColumn.js
