+++ ui/src/app/machines/views/MachineList/MachineListTable/DisksColumn/index.js
