+++ ui/src/app/machines/views/MachineList/MachineListTable/FabricColumn/FabricColumn.test.js
