+++ ui/src/app/machines/views/MachineList/MachineListTable/FabricColumn/index.js
