+++ ui/src/app/machines/views/MachineList/MachineListTable/NameColumn/NameColumn.js
