+++ ui/src/app/machines/views/MachineList/MachineListTable/NameColumn/index.js
