+++ ui/src/app/machines/views/MachineList/MachineListTable/PowerColumn/PowerColumn.js
