+++ ui/src/app/machines/views/MachineList/MachineListTable/PowerColumn/index.js
