+++ ui/src/app/machines/views/MachineList/MachineListTable/RamColumn/RamColumn.js
