+++ ui/src/app/machines/views/MachineList/MachineListTable/RamColumn/index.js
