+++ ui/src/app/machines/views/MachineList/MachineListTable/StatusColumn/StatusColumn.js
