+++ ui/src/app/machines/views/MachineList/MachineListTable/StatusColumn/index.js
