+++ ui/src/app/machines/views/MachineList/MachineListTable/StorageColumn/StorageColumn.js
