+++ ui/src/app/machines/views/MachineList/MachineListTable/StorageColumn/index.js
