+++ ui/src/app/machines/views/MachineList/MachineListTable/ZoneColumn/ZoneColumn.test.js
