+++ ui/src/app/machines/views/MachineList/MachineListTable/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MachineListTable";
